refactor and reorganize.
