+++ client/src/internals.rs
@@ -1,0 +1,6 @@
+// Copyright 2020 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod provider;
+
+pub use self::provider::Provider;
