+++ client/src/snapshot.rs
@@ -39,7 +39,7 @@ where {
     }
 
     /// Reads the data from the specified `&PathBuf` when given a `&str` password.  Returns a new `Snapshot`.
-    pub fn read_from_snapshot<P>(snapshot: &PathBuf, pass: &str) -> Self
+    pub fn read_from_snapshot<P>(snapshot: &PathBuf, pass: Vec<u8>) -> Self
     where
         P: BoxProvider + Clone + Send + Sync,
     {
@@ -48,13 +48,13 @@ where {
             .read(true)
             .open(snapshot)
             .expect("Unable to access snapshot. Make sure that it exists or run encrypt to build a new one.");
-        decrypt_snapshot(&mut file, &mut buffer, pass.as_bytes()).expect("unable to decrypt the snapshot");
+        // decrypt_snapshot(&mut file, &mut buffer, pass.as_bytes()).expect("unable to decrypt the snapshot");
 
         Snapshot::new::<P>(buffer)
     }
 
     /// Writes the data to the specified `&PathBuf` when given a `&str` password creating a new snapshot file.
-    pub fn write_to_snapshot(self, snapshot: &PathBuf, pass: &str) {
+    pub fn write_to_snapshot(self, snapshot: &PathBuf, pass: Vec<u8>) {
         let mut file = OpenOptions::new()
             .write(true)
             .create(true)
@@ -62,6 +62,6 @@ where {
             .expect("Unable to access snapshot. Make sure that it exists or run encrypt to build a new one.");
         // clear contents of the file before writing.
         file.set_len(0).expect("unable to clear the contents of the file file");
-        encrypt_snapshot(self.state, &mut file, pass.as_bytes()).expect("Couldn't write to the snapshot");
+        // encrypt_snapshot(self.state, &mut file, pass.as_bytes()).expect("Couldn't write to the snapshot");
     }
 }
