+++ client/src/utils.rs
@@ -1,14 +1,14 @@
-pub enum StatusMessage {
-    Ok,
-    Busy,
-    Error,
-}
+mod ask;
+#[allow(non_snake_case, dead_code)]
+mod hd;
+mod ids;
+mod types;
 
-pub enum StrongholdFlags {
-    Readable(bool),
-}
-
-pub enum VaultFlags {}
+pub use self::{
+    ask::ask,
+    ids::{ClientId, VaultId},
+    types::{StatusMessage, StrongholdFlags},
+};
 
 pub fn index_of_unchecked<T>(slice: &[T], item: &T) -> usize {
     if ::std::mem::size_of::<T>() == 0 {
